Add code for templates on studio.boardgamearena + reorganisation of folders
