+++ src/css/studio.css
@@ -1,0 +1,116 @@
+div.bgext_studio_row {
+  display: flex;
+  flex-flow: column;
+  gap: 0.3em;
+  width: 100%;
+  border: 1px solid black;
+  padding: 5px;
+  box-sizing: border-box;
+  margin: 5px;
+}
+
+div.bgext_studio_container {
+  display: flex;
+  flex-flow: row;
+  gap: 0.5em;
+  padding: 5px;
+  width: 700px;
+}
+
+div.bgext_studio_big_container {
+  display: flex;
+  flex-flow: row;
+  gap: 0.5em;
+  padding: 5px;
+  width: 100%;
+  box-sizing: border-box;
+}
+
+div.bgext_studio_select_container {
+  padding: 6px 12px 0px 6px;
+  flex-grow: 1;
+  margin: 4px 0;
+}
+
+div.bgext_studio_select_container>select {
+  width: 100%;
+  height: 32px;
+}
+
+div.bgext_studio_sentences {
+  border: 1px solid black;
+  display: flex;
+  flex-flow: column nowrap;
+  flex-grow: 1;
+}
+
+div.bgext_studio_sentence {
+  display: flex;
+  flex-flow: row nowrap;
+}
+
+div.bgext_studio_sentence_name {
+  cursor: pointer;
+  padding: 0.5em;
+  width: 150px;
+}
+
+div.bgext_studio_sentence_name>input {
+  width: 100%;
+  padding: 0.5em;
+}
+
+div.bgext_studio_sentence_text {
+  flex-grow: 1;
+  cursor: pointer;
+  padding: 0.5em;
+}
+
+div.bgext_studio_sentence_text>textarea {
+  width: 100%;
+  padding: 0.5em;
+  resize: none;
+  height: 100px;
+}
+
+div.bgext_studio_sentence_game {
+  position: relative;
+  display: flex;
+  flex-flow: row nowrap;
+  gap: 0.3em;
+  padding-left: 20px;
+}
+
+div.bgext_studio_sentence_game>input {
+  position: absolute;
+  top: 16px;
+  left: 0px;
+}
+
+div.bgext_studio_sentence_game>label {
+  padding: 1.1em 0.3em 0em 0em;
+}
+
+div.bgext_studio_sentence_action {
+  padding: 0.5em;
+}
+
+div.bgext_studio_sentence_hor_buts {
+  display: flex;
+  flex-flow: row;
+  gap: 1em;
+  justify-content: flex-end;
+  padding-right: 2em;
+}
+
+div.bgext_studio_sentence_col_title {
+  font-weight: bold;
+  line-height: 32PX;
+}
+
+div.bgext_studio_no_template {
+  width: 100%;
+  text-align: right;
+  font-style: italic;
+  line-height: 32px;
+}
