+++ src/js/ui/content/leftMenu/Avatar.jsx
