+++ src/js/ui/content/leftMenu/SideMenuItem.jsx
