+++ src/js/ui/content/leftMenu/icons/BoardIcon.jsx
