+++ src/js/ui/content/leftMenu/icons/CloseIcon.jsx
