+++ src/js/ui/content/leftMenu/icons/SandwichIcon.jsx
