+++ src/js/ui/content/leftMenu/icons/TopArrowIcon.jsx
